Move dss.events.handlers.index to dss.index.indexer
