+++ tests/test_subscriptions.py
@@ -17,11 +17,11 @@ pkg_root = os.path.abspath(os.path.join(os.path.dirname(__file__), '..'))  # noq
 sys.path.insert(0, pkg_root)  # noqa
 
 import dss
-from dss.events.handlers.index import BundleDocument
-from dss.logging import configure_test_logging
-from dss.index.es.manager import IndexManager
-from dss.util import UrlBuilder
 from dss.index.es import ElasticsearchClient
+from dss.index.es.document import BundleDocument
+from dss.index.es.manager import IndexManager
+from dss.logging import configure_test_logging
+from dss.util import UrlBuilder
 from tests import get_auth_header, get_bundle_fqid
 from tests.infra import DSSAssertMixin, testmode
 from tests.infra.elasticsearch_test_case import ElasticsearchTestCase
